pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fasthtml_materials"
version = "0.1.0"
description = "A simple CRUD app using FastHTML and PostgreSQL"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "python-fasthtml",
    "psycopg2-binary",
    "sqlalchemy",
    "alembic",
    "pytest",
    "pytest-asyncio",
    "pydantic",
]
